Dark Mode update
This commit creates the dark mode feature for the websites: 28M, career, error, home, music, support and tdm (en, ca, es). It also corrects some error and adds other little things to the home and career websites. Also the tdm websites got new font-size and a new css file for this website was created.
This commit also creates a /tdm/index.html file to test if this websites automatically pulls-up when searching for: "htpps:/joanalnu.github.io/tdm/" (if this will work cannot be proved in VSC preview).